Generating HTML-file

To generate an HTML table, use the show_table function. Next, we will show how you can customize the style and formatting of such a table.

TableView.show_table — Function

show_table(table::NTuple; kw...)

Generate html file based on AgGrid. The table passed to the function must be a NTuple whose elements are NamedTuple objects. Thus, each element of the table will represent its row.

Keyword arguments

column_settings::Dict = Dict(): Configuration for table filters. There you can specify following settings:

Column names as a string key and their parameters using another dictionary, which may contain the following pairs:

"filter" => ...: This parameter specifies how the current column should be filtered.

"text": Filter by text using list of values.

"number": Filter by number using range of values.

"date": Filter by date using range of values.

"formatter" => Dict(...): Here you can specify the desired formatting style of the columns.

"short" => true: Enables short form of the large numbers, like 4K (4,000) or 23M (23,000,000). By deafult is false.

"style" => ...: Pre-built styles for displaying numeric values.

"decimal": Default, plain number formatting.

"currency": Displays numeric values with corresponding currency symbol. Default used "USD" currency symbol, but you can change it by specifying a different currency "currency" => ... like "EUR", "CNY", etc. See the Current currency & funds code list.

"percent": Turn numeric values from interval [0.0, 1.0] to 0-100% percents.

"minimumFractionDigits" => ...: Specifies the minimum number of decimal places to display. Default is 0.

"maximumFractionDigits" => ...: Specifies the maximum number of decimal places to display. Default is 3.

"separator" => true: Enables grouping separators, such as thousands separators or thousand/lakh/crore separators.

"style" => Dict(...): Here, you can specify columns style

"text-align" => ...: Specifies the horizontal alignment of text in cell. Can be "left", "right" or "center".

"color" => ...: Change color of the text in cells. May be specifying by names of color ("red", "blue", etc.), hex codes ("#FFFFEE", "#3d85c6", etc.) or RGB values ("rgb(230, 230, 250)", "rgb(65, 10, 178)", etc.).

"background" => ...: Setting up the text background color.

"equals" => ...: Enables highlighting of the specified numeric value by color.

"threshold" => ...: Enables a separating border for highlight values that are greater or less than it. Colors of these values must be specified by "colorUp" or/and "ColorDown" parameters.

"colorUp" => ...: Color of the text that is less than "threshold" value.

"colorDown" => ...: Color of the text that is greater than "threshold" value.

Here, you can also specify a pair "cols" => true to enable the column filter on the right sidebar.

resize::Bool = true: Determines whether the column width can be changed by less than 150px.

out_file::String = "./index.html": Output file name.

Examples

julia> data = ((a = 0, b = 1),(a = 1, b = 35));

julia> settings = Dict("a" => Dict("filter" => "number"));

julia> TableView.showTable(data, column_settings=settings)
"./index.html"

Extended examples

Filter settings

headers = (:text_col, :number_col, :date_col)
table = (
    NamedTuple{headers}(("a", 1, "2024-01-01T00:00:00")),
    NamedTuple{headers}(("b", 2, "2024-02-01T00:00:00")),
    NamedTuple{headers}(("c", 3, "2024-03-01T00:00:00")),
)

settings = Dict(
    "text_col" => Dict(
        "filter" => "text"
    ),
    "number_col" => Dict(
        "filter" => "number"
    ),
    "date_col" => Dict(
        "filter" => "date"
    )
)

show_table(table, column_settings = settings, resize=true, out_file="./filter_example.html")

"./filter_example.html"

Formatter settings

headers = (:default_col, :short_col, :decimal_col, :usd_col, :eur_col, :percent_col, :separator_col)
table = (
    NamedTuple{headers}((1.0e6, 1.0e6, 1.0e6, 1.0e6, 1.0e6, 1.0e6, 1.0e6)),
    NamedTuple{headers}((1.0e3, 1.0e3, 1.0e3, 1.0e3, 1.0e3, 1.0e3, 1.0e3)),
    NamedTuple{headers}((1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0)),
    NamedTuple{headers}((1.0e-2, 1.0e-2, 1.0e-2, 1.0e-2, 1.0e-2, 1.0e-2, 1.0e-2)),
)

settings = Dict(
    "short_col" => Dict(
        "formatter" => Dict(
            "short" => true,
        )
    ),
    "decimal_col" => Dict(
        "formatter" => Dict(
            "style" => "decimal",
        )
    ),
    "usd_col" => Dict(
        "formatter" => Dict(
            "style" => "currency",
            "currency" => "USD",
        )
    ),
    "eur_col" => Dict(
        "formatter" => Dict(
            "style" => "currency",
            "currency" => "EUR",
        )
    ),
    "percent_col" => Dict(
        "formatter" => Dict(
            "style" => "percent",
        )
    ),
    "separator_col" => Dict(
        "formatter" => Dict(
            "separator" => true,
        )
    ),
)

show_table(table, column_settings = settings, resize=false, out_file="./formatter_example.html")

"./formatter_example.html"

Style settings

headers = (:color_col, :equals_col, :threshold_col)
table = (
    NamedTuple{headers}((1, 1, 1)),
    NamedTuple{headers}((2, 2, 2)),
    NamedTuple{headers}((3, 3, 3)),
    NamedTuple{headers}((4, 4, 4)),
    NamedTuple{headers}((5, 5, 5)),
)

settings = Dict(
    "color_col" => Dict(
        "style" => Dict(
            "color" => "cyan",
            "background" => "#34abeb",
            "text-align" => "left",
        )
    ),
    "equals_col" => Dict(
        "style" => Dict(
            "color" => "rgb(250, 100, 100)",
            "equals" => 3,
            "text-align" => "center",
        )
    ),
    "threshold_col" => Dict(
        "style" => Dict(
            "threshold" => 3.0,
            "colorUp" =>  "green",
            "colorDown" => "red",
        )
    ),
)

show_table(table, column_settings = settings, resize=false, out_file="./style_example.html")

"./style_example.html"